+++ examples/topic/cdc-cache-bus-freeseats/webserver.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/ydb-platform/ydb-go-sdk/v3"
 	"github.com/ydb-platform/ydb-go-sdk/v3/table"
-	"github.com/ydb-platform/ydb-go-sdk/v3/table/types"
+	"github.com/ydb-platform/ydb-go-sdk/v3/types"
 )
 
 var errNotEnthoughtFreeSeats = errors.New("not enough free seats")