+++ options.go
@@ -214,11 +214,11 @@ func WithConnectionString(connectionString string) Option {
 	}
 }
 
-// WithConnectionTTL defines duration for parking idle connections
+// Deprecated: connection parking was removed; this option is now a no-op.
+// Will be removed after Dec 2026.
+// Read about versioning policy: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#deprecated
 func WithConnectionTTL(ttl time.Duration) Option {
-	return func(ctx context.Context, d *Driver) error {
-		d.options = append(d.options, config.WithConnectionTTL(ttl))
-
+	return func(context.Context, *Driver) error {
 		return nil
 	}
 }
@@ -530,6 +530,7 @@ func WithQueryConfigOption(option queryConfig.Option) Option {
 }
 
 // WithSessionPoolSizeLimit set max size of internal sessions pool in table.Client
+// If sizeLimit is less than or equal to zero then the default pool size limit is used.
 func WithSessionPoolSizeLimit(sizeLimit int) Option {
 	return func(ctx context.Context, d *Driver) error {
 		d.tableOptions = append(d.tableOptions, tableConfig.WithSizeLimit(sizeLimit))